# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyPA",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/onboarding/goals": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Submit Onboarding Goals",
        "description": "Zero items in both lists is valid (onboarding is optional per-item) — still marks\n`completed`. `memory_record_id` stays `NULL` for every suggestion-derived goal —\nsynthesis spans many extraction records, not one FK-able one.\n\nA checked suggestion the user already has a goal for is not created again (its open\nmatch gets its blanks filled instead — core/goals/onboarding.py), so the response lists\nonly newly created goals. Typed `additional_goals` are not deduped: like `POST /goals`,\nthe user typed them on purpose.",
        "operationId": "submit_onboarding_goals_api_v1_onboarding_goals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingGoalsSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GoalRead"
                  },
                  "type": "array",
                  "title": "Response Submit Onboarding Goals Api V1 Onboarding Goals Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/v1/goals": {
      "post": {
        "tags": [
          "goals"
        ],
        "summary": "Write Goal",
        "operationId": "write_goal_api_v1_goals_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "Read Goals",
        "description": "`status` takes several values (`?status=open&status=paused` for an \"active goals\"\nview). Newest first — the tasks list's \"open first\" trick relies on the alphabetical\norder of two statuses and doesn't work with four. Answers `304` when the browser's\ncached copy is still current (core/utils/etag.py).",
        "operationId": "read_goals_api_v1_goals_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "items_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Items Per Page"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "open",
                      "paused",
                      "done",
                      "dropped"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "short_term",
                    "long_term"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Horizon"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "manual",
                    "conversation",
                    "email",
                    "calendar",
                    "notion"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedListResponse_GoalRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}": {
      "get": {
        "tags": [
          "goals"
        ],
        "summary": "Read Goal",
        "operationId": "read_goal_api_v1_goals__goal_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "goals"
        ],
        "summary": "Patch Goal",
        "operationId": "patch_goal_api_v1_goals__goal_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "goals"
        ],
        "summary": "Erase Goal",
        "operationId": "erase_goal_api_v1_goals__goal_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalDeletedRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/goals/{goal_id}/status": {
      "patch": {
        "tags": [
          "goals"
        ],
        "summary": "Patch Goal Status",
        "operationId": "patch_goal_status_api_v1_goals__goal_id__status_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "goal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Goal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoalStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GoalCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "short_term",
                  "long_term"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "GoalCreate",
        "description": "Public input for `POST /goals`. `horizon: None` means \"left empty — let the AI guess\nin the background\" (core/goals/jobs.py). `source` is server-set (`\"manual\"`), never\nclient-supplied."
      },
      "GoalDeletedRead": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "GoalDeletedRead",
        "description": "Response shape for `DELETE /goals/{id}`."
      },
      "GoalRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "paused",
              "done",
              "dropped"
            ],
            "title": "Status"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "conversation",
              "email",
              "calendar",
              "notion"
            ],
            "title": "Source"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "short_term",
                  "long_term"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          },
          "memory_record_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Record Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "title_manually_set": {
            "type": "boolean",
            "title": "Title Manually Set",
            "default": false
          },
          "description_manually_set": {
            "type": "boolean",
            "title": "Description Manually Set",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "source",
          "created_at"
        ],
        "title": "GoalRead"
      },
      "GoalStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "paused",
              "done",
              "dropped"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "GoalStatusUpdate"
      },
      "GoalUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "short_term",
                  "long_term"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          },
          "target_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "GoalUpdate",
        "description": "Public input for `PATCH /goals/{id}`. `status` is deliberately absent — it changes\nonly through `PATCH /goals/{id}/status`, the one place that enqueues Notion sync. An\nexplicit `null` clears `description`/`horizon`/`target_date`, but is rejected for the\nnon-nullable `title`."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnboardingGoalsSubmitRequest": {
        "properties": {
          "checked_suggestions": {
            "items": {
              "$ref": "#/components/schemas/SuggestedGoal"
            },
            "type": "array",
            "title": "Checked Suggestions"
          },
          "additional_goals": {
            "items": {
              "$ref": "#/components/schemas/SuggestedGoal"
            },
            "type": "array",
            "title": "Additional Goals"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "OnboardingGoalsSubmitRequest"
      },
...
      "PaginatedListResponse_GoalRead_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/GoalRead"
            },
            "type": "array",
            "title": "Data"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "items_per_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items Per Page"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total_count",
          "has_more"
        ],
        "title": "PaginatedListResponse[GoalRead]"
      },
...
      "SuggestedGoal": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "horizon": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "short_term",
                  "long_term"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "SuggestedGoal",
        "description": "The one canonical shape for: the synthesis LLM's structured output, the JSONB\npersisted on `users.onboarding_suggested_goals`, and what the client posts to\n`POST /onboarding/goals`."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/login"
          }
        }
      }
    }
  }
}